separate subtasks_defs.py & subviews.py
